imap: split message functions in a separate file
